fixing static code warnings

+++ mindspore/core/ops/grad/adaptive_avg_pool_3d_grad.cc
@@ -52,7 +52,7 @@ abstract::ShapePtr InferShapeAdaptiveAvgPool3DGrad(const PrimitivePtr &primitive
       auto value = static_cast<int32_t *>(orig_input_shape_tensor->data_c());
       MS_EXCEPTION_IF_NULL(value);
       for (int64_t i = 0; i < input_grad_dims; ++i) {
-        orig_input_shape_value_vec[i] = value[i] > 0 ? static_cast<int64_t>(value[i]) : 1;
+        orig_input_shape_value_vec[i] = value[i] > 0 ? static_cast<int64_t>(value[i]) : static_cast<int64_t>(1);
       }
       gen_value_succ = true;
     }
